Rename object -> data
