+++ moq-transport/src/coding/decode.rs
@@ -48,8 +48,4 @@ pub enum DecodeError {
 
 	#[error("io error: {0}")]
 	IoError(#[from] std::io::Error),
-
-	// Used to signal that the stream has ended.
-	#[error("no more messages")]
-	Final,
 }
