+++ moq-transport/src/lib.rs
@@ -10,8 +10,8 @@ mod coding;
 mod error;
 
 pub mod cache;
+pub mod data;
 pub mod message;
-pub mod object;
 pub mod session;
 pub mod setup;
